+++ colpali_engine/evaluation/__init__.py
@@ -1,1 +1,0 @@
-from .retrieval_scorer import RetrievalScorer