+++ colpali_engine/models/paligemma/colpali/processing_colpali.py
@@ -1,5 +1,6 @@
-from typing import List, Optional
+from typing import List, Optional, Union
 
+import torch
 from PIL import Image
 from transformers import BatchFeature, PaliGemmaProcessor
 
@@ -62,7 +63,24 @@ class ColPaliProcessor(BaseVisualRetrieverProcessor, PaliGemmaProcessor):
 
         del batch_query["pixel_values"]
 
-        batch_query["input_ids"] = batch_query["input_ids"][..., self.image_seq_length :]
-        batch_query["attention_mask"] = batch_query["attention_mask"][..., self.image_seq_length :]
+        batch_query["input_ids"] = batch_query["input_ids"][
+            ..., self.image_seq_length :
+        ]
+        batch_query["attention_mask"] = batch_query["attention_mask"][
+            ..., self.image_seq_length :
+        ]
 
         return batch_query
+
+    def score(
+            self,
+            qs: List[torch.Tensor],
+            ps: List[torch.Tensor],
+            batch_size: int = 128,
+            device: Union[Optional[str], torch.device] = None,
+    ) -> torch.Tensor:
+        """
+        Compute the MaxSim score (ColBERT-like) for the given multi-vector query and passage embeddings.
+        """
+
+        return self.score_multi_vector(qs, ps, batch_size=batch_size, device=device)