+++ tests/evaluation/test_retrieval_scorer.py
@@ -1,28 +1,26 @@
+from unittest.mock import patch
+
 import pytest
 import torch
 
-from colpali_engine.evaluation.retrieval_scorer import RetrievalScorer
+from colpali_engine.utils.processing_utils import BaseVisualRetrieverProcessor
 
 
 @pytest.fixture
-def single_vector_scorer() -> RetrievalScorer:
-    return RetrievalScorer(is_multi_vector=False, device="cpu")
+@patch.multiple(BaseVisualRetrieverProcessor, __abstractmethods__=set())
+def vector_scorer() -> BaseVisualRetrieverProcessor:
+    return BaseVisualRetrieverProcessor()
 
 
-@pytest.fixture
-def multi_vector_scorer() -> RetrievalScorer:
-    return RetrievalScorer(is_multi_vector=True, device="cpu")
-
-
-def test_score_single_vector_embeddings(single_vector_scorer: RetrievalScorer):
+def test_score_single_vector_embeddings(vector_scorer: BaseVisualRetrieverProcessor):
     qs = [torch.randn(128) for _ in range(10)]
     ps = [torch.randn(128) for _ in range(10)]
-    scores = single_vector_scorer.evaluate(qs, ps)
+    scores = vector_scorer.score_single_vector(qs, ps)
     assert scores.shape == (len(qs), len(ps))
 
 
-def test_score_multi_vector_embeddings(multi_vector_scorer: RetrievalScorer):
+def test_score_multi_vector_embeddings(vector_scorer: BaseVisualRetrieverProcessor):
     qs = [torch.randn(10, 128) for _ in range(10)]
     ps = [torch.randn(42, 128) for _ in range(10)]
-    scores = multi_vector_scorer.evaluate(qs, ps)
+    scores = vector_scorer.score_multi_vector(qs, ps)
     assert scores.shape == (len(qs), len(ps))